[IMPROVE][ENTERPRISE] Improve how micro services are loaded (#24388)
