+++ app/search/server/search.internalService.ts
@@ -1,11 +1,11 @@
 import { Users } from '../../models/server';
 import { settings } from '../../settings/server';
 import { searchProviderService } from './service/providerService';
-import { ServiceClass } from '../../../server/sdk/types/ServiceClass';
+import { ServiceClassInternal } from '../../../server/sdk/types/ServiceClass';
 import { api } from '../../../server/sdk/api';
 import { searchEventService } from './events/events';
 
-class Search extends ServiceClass {
+class Search extends ServiceClassInternal {
 	protected name = 'search';
 
 	protected internal = true;
