+++ app/tokenpass/server/roomAccessValidator.internalService.ts
@@ -1,11 +1,10 @@
-import { ServiceClass } from '../../../server/sdk/types/ServiceClass';
+import { ServiceClassInternal } from '../../../server/sdk/types/ServiceClass';
 import { validators } from './roomAccessValidator.compatibility';
-import { api } from '../../../server/sdk/api';
 import { IAuthorizationTokenpass } from '../../../server/sdk/types/IAuthorizationTokenpass';
 import { IRoom } from '../../../definition/IRoom';
 import { IUser } from '../../../definition/IUser';
 
-class AuthorizationTokenpass extends ServiceClass implements IAuthorizationTokenpass {
+export class AuthorizationTokenpass extends ServiceClassInternal implements IAuthorizationTokenpass {
 	protected name = 'authorization-tokenpass';
 
 	protected internal = true;
@@ -20,5 +19,3 @@ class AuthorizationTokenpass extends ServiceClass implements IAuthorizationToken
 		return false;
 	}
 }
-
-api.registerService(new AuthorizationTokenpass());
