+++ ee/app/license/server/index.ts
@@ -1,4 +1,3 @@
-import './license.internalService';
 import './settings';
 import './methods';
 import './startup';
