+++ ee/app/license/server/license.internalService.ts
@@ -1,15 +1,14 @@
-import { ServiceClass } from '../../../../server/sdk/types/ServiceClass';
+import { debounce } from 'underscore';
+
 import { api } from '../../../../server/sdk/api';
 import { ILicense } from '../../../../server/sdk/types/ILicense';
-import { hasLicense, isEnterprise, getModules, onValidateLicenses, onModule } from './license';
-import { resetEnterprisePermissions } from '../../authorization/server/resetEnterprisePermissions';
-import { Authorization } from '../../../../server/sdk';
+import { ServiceClassInternal } from '../../../../server/sdk/types/ServiceClass';
 import { guestPermissions } from '../../authorization/lib/guestPermissions';
+import { resetEnterprisePermissions } from '../../authorization/server/resetEnterprisePermissions';
+import { getModules, hasLicense, isEnterprise, onModule, onValidateLicenses } from './license';
 
-class LicenseService extends ServiceClass implements ILicense {
+export class LicenseService extends ServiceClassInternal implements ILicense {
 	protected name = 'license';
-
-	protected internal = true;
 
 	constructor() {
 		super();
@@ -19,13 +18,55 @@ class LicenseService extends ServiceClass implements ILicense {
 				return;
 			}
 
-			Authorization.addRoleRestrictions('guest', guestPermissions);
+			api.broadcast('authorization.guestPermissions', guestPermissions);
 			resetEnterprisePermissions();
 		});
 
 		onModule((licenseModule) => {
 			api.broadcast('license.module', licenseModule);
 		});
+
+		/**
+		 * The main idea is if there is no scalability module enabled,
+		 * then we should not allow more than one service per environment.
+		 * So we list the services and nodes, and if there is more than
+		 * one, we inform the service that it should be disabled.
+		 */
+		this.onEvent(
+			'$services.changed',
+			debounce(async () => {
+				if (hasLicense('scalability')) {
+					return;
+				}
+
+				const services: {
+					name: string;
+					nodes: string[];
+				}[] = await api.call('$node.services');
+
+				// Filter only the duplicated services
+				const duplicated = services.filter((service) => {
+					return service.name !== '$node' && service.nodes.length > 1;
+				});
+
+				if (!duplicated.length) {
+					return;
+				}
+
+				const brokers: Record<string, string[]> = Object.fromEntries(
+					duplicated.map((service) => {
+						// remove the first node from the list
+						const [, ...nodes] = service.nodes;
+						return [service.name, nodes];
+					}),
+				);
+
+				const duplicatedServicesNames = duplicated.map((service) => service.name);
+
+				// send shutdown signal to the duplicated services
+				api.broadcastToServices(duplicatedServicesNames, 'shutdown', brokers);
+			}, 1000),
+		);
 	}
 
 	async started(): Promise<void> {
@@ -33,7 +74,7 @@ class LicenseService extends ServiceClass implements ILicense {
 			return;
 		}
 
-		Authorization.addRoleRestrictions('guest', guestPermissions);
+		api.broadcast('authorization.guestPermissions', guestPermissions);
 		resetEnterprisePermissions();
 	}
 
@@ -48,6 +89,8 @@ class LicenseService extends ServiceClass implements ILicense {
 	getModules(): string[] {
 		return getModules();
 	}
-}
 
-api.registerService(new LicenseService());
+	getGuestPermissions(): string[] {
+		return guestPermissions;
+	}
+}
