+++ ee/app/settings/server/settings.internalService.ts
@@ -1,10 +1,9 @@
-import { ServiceClass } from '../../../../server/sdk/types/ServiceClass';
-import { api } from '../../../../server/sdk/api';
+import { ServiceClassInternal } from '../../../../server/sdk/types/ServiceClass';
 import { IEnterpriseSettings } from '../../../../server/sdk/types/IEnterpriseSettings';
 import { changeSettingValue } from './settings';
 import { ISetting } from '../../../../definition/ISetting';
 
-class EnterpriseSettings extends ServiceClass implements IEnterpriseSettings {
+export class EnterpriseSettings extends ServiceClassInternal implements IEnterpriseSettings {
 	protected name = 'ee-settings';
 
 	protected internal = true;
@@ -13,5 +12,3 @@ class EnterpriseSettings extends ServiceClass implements IEnterpriseSettings {
 		return changeSettingValue(record);
 	}
 }
-
-api.registerService(new EnterpriseSettings());
