+++ ee/server/local-services/ldap/service.ts
@@ -1,11 +1,8 @@
-import '../../broker';
-
-import { api } from '../../../../server/sdk/api';
 import { LDAPEEManager } from '../../lib/ldap/Manager';
 import { ILDAPEEService } from '../../sdk/types/ILDAPEEService';
-import { ServiceClass } from '../../../../server/sdk/types/ServiceClass';
+import { ServiceClassInternal } from '../../../../server/sdk/types/ServiceClass';
 
-export class LDAPEEService extends ServiceClass implements ILDAPEEService {
+export class LDAPEEService extends ServiceClassInternal implements ILDAPEEService {
 	protected name = 'ldap-enterprise';
 
 	async sync(): Promise<void> {
@@ -20,5 +17,3 @@ export class LDAPEEService extends ServiceClass implements ILDAPEEService {
 		return LDAPEEManager.syncLogout();
 	}
 }
-
-api.registerService(new LDAPEEService());
