+++ ee/server/services/ddp-streamer/service.ts
@@ -1,4 +1,4 @@
-import '../../broker';
+import '../../startup/broker';
 
 import { api } from '../../../../server/sdk/api';
 import { DDPStreamer } from './DDPStreamer';
