+++ ee/server/services/stream-hub/service.ts
@@ -1,4 +1,4 @@
-import '../../broker';
+import '../../startup/broker';
 
 import { api } from '../../../../server/sdk/api';
 import { StreamHub } from './StreamHub';
