+++ ee/server/startup/index.ts
@@ -1,2 +1,10 @@
 import './engagementDashboard';
 import './seatsCap';
+import './services';
+
+const { TRANSPORTER = '' } = process.env;
+
+// only starts network broker if transporter properly configured
+if (TRANSPORTER.match(/^(?:nats|TCP)/)) {
+	require('./broker');
+}
