+++ server/sdk/types/IAuthorization.ts
@@ -11,7 +11,6 @@ export interface IAuthorization {
 	hasAllPermission(userId: string, permissions: string[], scope?: string): Promise<boolean>;
 	hasPermission(userId: string, permissionId: string, scope?: string): Promise<boolean>;
 	hasAtLeastOnePermission(userId: string, permissions: string[], scope?: string): Promise<boolean>;
-	addRoleRestrictions(role: string, permissions: string[]): Promise<void>;
 	canAccessRoom: RoomAccessValidator;
 	canAccessRoomId(rid: IRoom['_id'], uid: IUser['_id']): Promise<boolean>;
 }
