+++ server/services/image/service.ts
@@ -4,14 +4,14 @@ import fileType from 'file-type';
 import sharp from 'sharp';
 import isSvg from 'is-svg';
 
-import { ServiceClass } from '../../sdk/types/ServiceClass';
+import { ServiceClassInternal } from '../../sdk/types/ServiceClass';
 import { IMediaService, ResizeResult } from '../../sdk/types/IMediaService';
 
 /* eslint-disable  @typescript-eslint/no-var-requires */
 const ExifTransformer = require('exif-be-gone');
 /* eslint-enable  @typescript-eslint/no-var-requires */
 
-export class MediaService extends ServiceClass implements IMediaService {
+export class MediaService extends ServiceClassInternal implements IMediaService {
 	protected name = 'media';
 
 	private imageExts = new Set([
