+++ server/services/ldap/service.ts
@@ -1,9 +1,9 @@
 import { LDAPManager } from '../../lib/ldap/Manager';
 import { ILDAPService } from '../../sdk/types/ILDAPService';
-import { ServiceClass } from '../../sdk/types/ServiceClass';
+import { ServiceClassInternal } from '../../sdk/types/ServiceClass';
 import { LDAPLoginResult } from '../../../definition/ldap/ILDAPLoginResult';
 
-export class LDAPService extends ServiceClass implements ILDAPService {
+export class LDAPService extends ServiceClassInternal implements ILDAPService {
 	protected name = 'ldap';
 
 	async loginRequest(username: string, password: string): Promise<LDAPLoginResult> {
