+++ server/services/room/service.ts
@@ -1,6 +1,6 @@
 import { Db } from 'mongodb';
 
-import { ServiceClass } from '../../sdk/types/ServiceClass';
+import { ServiceClassInternal } from '../../sdk/types/ServiceClass';
 import { ICreateRoomParams, IRoomService } from '../../sdk/types/IRoomService';
 import { Authorization } from '../../sdk';
 import { IRoom } from '../../../definition/IRoom';
@@ -8,7 +8,7 @@ import { UsersRaw } from '../../../app/models/server/raw/Users';
 import { createRoom } from '../../../app/lib/server/functions/createRoom'; // TODO remove this import
 import { IUser } from '../../../definition/IUser';
 
-export class RoomService extends ServiceClass implements IRoomService {
+export class RoomService extends ServiceClassInternal implements IRoomService {
 	protected name = 'room';
 
 	private Users: UsersRaw;
